Move FloatingMenu template component
