+++ frontend/src/components/layout/FloatingMenu.vue
